libsnapshot: ensure snapuserd is in ublk mode
Introduces a helper function EnsureSnapuserdIsUblk to verify
and enforce that the snapuserd daemon is running in the correct
mode before attempting to map a ublk-based snapshot.

If snapuserd is found to be in dm-user mode, this function will
stop the daemon, write a hint file to force ublk mode
and then restart it. This prevents snapshot mapping
failures caused by a mismatched daemon configuration.

The MapUserspaceCowUblk function is updated to use this new
helper, making the snapshot creation process more robust.
Additionally, WaitForServiceToTerminate is made public to be
accessible from the SnapshotManager.

Bug: 444696988
Test: th, manual ublk based OTA on a device

Change-Id: I04b22147eda05ee91611c3e9d3c467cc861c5228
Signed-off-by: Sandeep Dhavale <dhavale@google.com>